+++ src/components/TokenList/TokenList.styles.tsx
@@ -2,10 +2,7 @@ import styled from "styled-components/macro";
 
 import isActiveLanguageLogographic from "../../helpers/isActiveLanguageLogographic";
 import { InputTextStyle } from "../../style/mixins";
-import { fontWide } from "../../style/themes";
 import Icon from "../Icon/Icon";
-import LoadingSpinner from "../LoadingSpinner/LoadingSpinner";
-import { SpinningIcon } from "../LoadingSpinner/LoadingSpinner.styles";
 import TextInput from "../TextInput/TextInput";
 import { StyledInput } from "../TextInput/TextInput.styles";
 
@@ -44,49 +41,6 @@ export const SearchInput = styled(TextInput)`
   }
 `;
 
-export const TokensScrollContainer = styled.div`
-  position: relative;
-  width: 100%;
-  height: fit-content;
-  min-height: 10rem;
-`;
-
-export const TokensContainer = styled.div``;
-
-export const TokenListLoader = styled(LoadingSpinner)`
-  &::before {
-    content: "";
-    position: absolute;
-    inset: 0;
-    background: ${(props) => props.theme.colors.darkBlue};
-    opacity: 0.35;
-  }
-
-  position: absolute;
-  inset: 0;
-
-  ${SpinningIcon} {
-    width: 2rem;
-    height: 2rem;
-  }
-`;
-
-export const Legend = styled.div`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  min-height: 1.5rem;
-  margin-block-start: 2rem;
-`;
-
-export const LegendItem = styled.div`
-  word-break: keep-all;
-  font-family: ${fontWide};
-  font-weight: 500;
-  font-size: ${() => (isActiveLanguageLogographic() ? "0.875rem" : "1rem")};
-  color: ${(props) => props.theme.colors.lightGrey};
-`;
-
 export const InactiveTitleContainer = styled.div`
   border: 1px solid ${(props) => props.theme.colors.borderGrey};
   border-radius: 2px;
@@ -108,7 +62,4 @@ export const InformationIcon = styled(Icon)`
   margin-left: 0.25rem;
 `;
 
-export const NoResultsContainer = styled.div`
-  margin-block-start: 1rem;
-  text-align: center;
-`;
+export const TokensContainer = styled.div``;
